+++ AutomatedTesting/Gem/PythonTests/EditorPythonTestTools/editor_python_test_tools/pyside_component_utils.py
@@ -7,7 +7,7 @@ SPDX-License-Identifier: Apache-2.0 OR MIT
 
 import PySide2
 
-import editor_python_test_tools.pyside_utils as pyside_utils
+import pyside_utils
 
 def get_component_combobox_values(component_name, property_name, log_fn=None):
     """
