+++ AutomatedTesting/Gem/PythonTests/Prefab/tests/detach_prefab/DetachPrefab_UnderAnotherPrefab.py
@@ -11,7 +11,7 @@ def DetachPrefab_UnderAnotherPrefab():
     CAR_PREFAB_FILE_NAME = Path(__file__).stem + 'car_prefab'
     WHEEL_PREFAB_FILE_NAME = Path(__file__).stem + 'wheel_prefab'
 
-    import editor_python_test_tools.pyside_utils as pyside_utils
+    import pyside_utils
 
     @pyside_utils.wrap_async
     async def run_test():
