+++ AutomatedTesting/Gem/PythonTests/Prefab/tests/reparent_prefab/ReparentPrefab_UnderPrefabAndEntityHierarchies.py
@@ -12,7 +12,7 @@ def ReparentPrefab_UnderPrefabAndEntityHierarchies():
     WHEEL_PREFAB_FILE_NAME = Path(__file__).stem + '_wheel_prefab'
     DRIVER_PREFAB_FILE_NAME = Path(__file__).stem + '_driver_prefab'
 
-    import editor_python_test_tools.pyside_utils as pyside_utils
+    import pyside_utils
 
     @pyside_utils.wrap_async
     async def run_test():
