+++ AutomatedTesting/Gem/PythonTests/scripting/ScriptEvents_Default_SendReceiveSuccessfully.py
@@ -6,7 +6,7 @@ SPDX-License-Identifier: Apache-2.0 OR MIT
 """
 
 import os
-import editor_python_test_tools.pyside_utils as pyside_utils
+import pyside_utils
 from editor_python_test_tools.utils import Report, Tracer
 from editor_python_test_tools.utils import TestHelper as helper
 import editor_python_test_tools.hydra_editor_utils as hydra
